Added try/except to get rid of deprecation warning
The issue is that self.X is an empty list, [ ], until it is populated at which time it becomes a numpy array. Trying as a numpy array, then as an exception trying as a list works, but it feels janky. Because the list is never used as a list I am going to try changing self.X to initialize as a size 0 np array.