Modified bessy import statements to use relative imports
Removed use of `sys.path.append` to use local files, and changed import calls of bci_essentials (bessy python) to use relative import statements.

Updated both the core bci_essentials code (which was all within the classification submodule) and all the examples.

Also updated the smoke test to do the same.

Also updated the filenames within the examples to do relative paths to the data in a way that is agnostic to the OS used (i.e. removed us of hardcoded back-slash or forward-slash)